requirements.txt
streamlit==0.89.0
google-generativeai==0.1.0